backend/requirement.txt
flask
flask-cors
numpy
onnxruntime
xgboost
joblib
pillow
